Added H1 to indexes
