+++ MacroPepelelipa/Database/Model/NotebookObject.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2020 Pedro Giuliano Farina. All rights reserved.
 //
 
+import UIKit
+
 internal class NotebookObject: NotebookEntity {
 
     private weak var workspace: WorkspaceObject?
@@ -38,7 +40,12 @@ internal class NotebookObject: NotebookEntity {
         var indexes: [NotebookIndexObject] = []
         for note in notes {
             indexes.append(NotebookIndexObject(index: note.title.string, note: note, isTitle: true))
-            //TODO: get the H1 texts from note
+            note.text.enumerateAttribute(.font, in: NSRange(location: 0, length: note.text.length), options: .longestEffectiveRangeNotRequired) { (font, range, _) in
+                if font as? UIFont == UIFont.firstHeaderFont {
+                    let text = note.text.attributedSubstring(from: range).string
+                    indexes.append(NotebookIndexObject(index: text, note: note, isTitle: false))
+                }
+            }
         }
         return indexes
     }
